requirements.txt
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
certifi==2015.11.20.1
py==1.4.31
pycurl==7.19.5.3
pytest==2.8.3
PyYAML==3.11
singledispatch==3.4.0.3
six==1.10.0
tornado==4.3
ujson==1.34
wheel==0.26.0